:construction: Splitage in different channels
