+++ .env.example
@@ -2,7 +2,9 @@ TOKEN=
 PREFIX=
 OWNER_ID=
 IDDISCORDHEAVEN=
-IDCHANNELHEAVENBOT=
+IDCHANNELEVAN=
+IDCHANNELDONJON=
+IDCHANNELGAZETTE=
 EMOJITANK=U+1f6e1
 EMOJIHEAL=U+1f489
 EMOJIDPS=U+2694
