+++ src/main/java/be/isservers/hmb/CommandManager.java
@@ -6,7 +6,7 @@ import be.isservers.hmb.command.privateCommand.LfgCommand;
 import be.isservers.hmb.command.publicCommands.HelpCommand;
 import be.isservers.hmb.command.publicCommands.JokeCommand;
 import be.isservers.hmb.command.publicCommands.MemeCommand;
-import be.isservers.hmb.command.publicCommands.PingCommand;
+import be.isservers.hmb.command.publicCommands.admin.PingCommand;
 import be.isservers.hmb.command.publicCommands.admin.ClearChannelCommand;
 import be.isservers.hmb.command.publicCommands.admin.SetPrefixCommand;
 import be.isservers.hmb.command.publicCommands.music.*;
