+++ src/main/java/be/isservers/hmb/Listener.java
@@ -36,7 +36,7 @@ public class Listener extends ListenerAdapter {
             for (Guild guild : event.getJDA().getGuilds()){
                 if(guild.getId().equals(Config.getIdDiscordHeaven())){
                     LOGGER.info("Connected to " + guild.getName(), event.getJDA().getSelfUser().getAsTag());
-                    LFGmain.Clear(guild.getTextChannelById(Config.getIdChannelHeavenBot()));
+                    LFGmain.Clear(guild.getTextChannelById(Config.getIdChannelDonjon()));
                     LFGdataManagement.heavenDiscord = guild;
                     LFGdataManagement.InitializeOrganizedDate(event);
                     //new Timer(TIME_BETWEEN_AUTO_DELETE,new LFGautoDeleteEvent()).start();
